refactor(sybra): split init and completion
Final step of issue 604. App.go reaches 391 LOC (was 1382 at
start of refactor), holding only: struct, NewApp, Options,
Startup, Shutdown, Context, and the 8 Wails-bound entry points.

Two new files in internal/sybra:

- agent_completion.go: AgentCompletionHandler embeds DomainHandler
  and owns OnComplete + OnWorkflowComplete + recordRunStats.
  Constructed in wireServices once every dep is set, then
  subscribed via agents.SetOnComplete and engine.SetOnComplete.

- app_init.go: 20 init helpers (initAudit, initStats, initBgops,
  initFileWatcher, initStatusHook, initProviderHealth,
  initWorkflowEngine, initApprovalServer, initLoopAgents,
  initLoopScheduler, seedDefaultLoopAgents, initIssuesFetcher,
  initAutomations, emitDegradedWarnings, logAutomationsSummary,
  allowsProjectType, logAudit, newRecovery, syncSkillsBundle)
  plus MonitorReportBinding type. Methods stay on *App so test
  fixtures and callers (lifecycle.go, app_renovate.go,
  app_automation_routing_test.go) keep working.

The TestOnAgentComplete_EmptyTaskID_NoCrash test instantiates
AgentCompletionHandler directly instead of going through *App.

Step 3/3 of issue 604.
