+++ internal/sybra/app_init.go
@@ -1,0 +1,353 @@
+package sybra
+
+import (
+	"context"
+	"os"
+	"path/filepath"
+	"time"
+
+	"github.com/Automaat/sybra/internal/agent"
+	"github.com/Automaat/sybra/internal/audit"
+	"github.com/Automaat/sybra/internal/bgop"
+	"github.com/Automaat/sybra/internal/config"
+	"github.com/Automaat/sybra/internal/events"
+	"github.com/Automaat/sybra/internal/loopagent"
+	"github.com/Automaat/sybra/internal/monitor"
+	"github.com/Automaat/sybra/internal/notification"
+	"github.com/Automaat/sybra/internal/poll"
+	"github.com/Automaat/sybra/internal/project"
+	"github.com/Automaat/sybra/internal/provider"
+	"github.com/Automaat/sybra/internal/recovery"
+	"github.com/Automaat/sybra/internal/skillsync"
+	"github.com/Automaat/sybra/internal/stats"
+	"github.com/Automaat/sybra/internal/task"
+	"github.com/Automaat/sybra/internal/watcher"
+	"github.com/Automaat/sybra/internal/workflow"
+)
+
+func (a *App) initBgops(emit func(string, any)) {
+	a.bgops = bgop.NewTracker(emit, filepath.Join(config.HomeDir(), "bgops.json"))
+	a.bgops.LoadFromDisk()
+}
+
+func (a *App) initFileWatcher(ctx context.Context, emit func(string, any)) {
+	w := watcher.New(a.tasksDir, emit, a.logger)
+	a.watcher = w
+	if err := w.Start(ctx); err != nil {
+		a.logger.Error("watcher.start", "err", err)
+	}
+}
+
+// MonitorReportBinding is the Wails-friendly envelope for the latest
+// monitor report. Keeping the struct here (rather than in internal/monitor)
+// avoids the frontend bindings needing to handle a `monitor.Report | null`
+// union — Enabled/Ready flags say whether Report is populated.
+type MonitorReportBinding struct {
+	Enabled bool           `json:"enabled"`
+	Ready   bool           `json:"ready"`
+	Report  monitor.Report `json:"report"`
+}
+
+// allowsProjectType reports whether project-scoped automations on this machine
+// should act on the given project type. Used to route automation work between
+// instances (e.g., pet projects on the server, work projects on the laptop).
+func (a *App) allowsProjectType(t project.ProjectType) bool {
+	return a.cfg.AllowsProjectType(string(t))
+}
+
+// initIssuesFetcher constructs the GitHub Issues fetcher if enabled, returning
+// nil otherwise. Kept separate so Startup stays under the funlen limit.
+func (a *App) initIssuesFetcher(emit func(string, any)) *poll.IssuesFetcher {
+	if !a.cfg.GitHub.Enabled {
+		a.logger.Info("github.disabled")
+		return nil
+	}
+	return poll.NewIssuesFetcher(a.tasks, a.projects, emit, a.logger, a.allowsProjectType)
+}
+
+// logAutomationsSummary logs a one-line snapshot of which automations this
+// machine runs. Useful when comparing two instances side by side.
+func (a *App) logAutomationsSummary() {
+	loopAgentsEnabled := 0
+	if a.loopAgents != nil {
+		if las, err := a.loopAgents.List(); err == nil {
+			for i := range las {
+				if las[i].Enabled {
+					loopAgentsEnabled++
+				}
+			}
+		}
+	}
+	projectTypes := a.cfg.ProjectTypes
+	if len(projectTypes) == 0 {
+		projectTypes = []string{"*"}
+	}
+	a.logger.Info("app.automations",
+		"todoist", a.cfg.Todoist.Enabled && a.cfg.Todoist.APIToken != "",
+		"github", a.cfg.GitHub.Enabled,
+		"renovate", a.cfg.Renovate.Enabled,
+		"triage", a.cfg.Triage.Enabled,
+		"human_review", a.humanReview != nil,
+		"project_types", projectTypes,
+		"loop_agents_enabled", loopAgentsEnabled,
+	)
+}
+
+func (a *App) initStats() {
+	statsStore, err := stats.NewStore(config.StatsFile())
+	if err != nil {
+		a.logger.Warn("stats.init.degraded", "err", err)
+		// a.stats remains nil; StatsService.GetStats() guards against nil.
+		return
+	}
+	a.stats = statsStore
+	if err := statsStore.Backfill(a.auditDir); err != nil {
+		a.logger.Warn("stats.backfill", "err", err)
+	}
+}
+
+func (a *App) initStatusHook() {
+	a.tasks.SetStatusChangeHook(func(taskID, from, to string) {
+		a.logAudit(audit.EventTaskStatusChanged, taskID, "", map[string]any{"from": from, "to": to})
+
+		// Advance workflows whose current run_agent step declares a
+		// matching wait_for_status. This is how interactive agents (which
+		// never exit between turns) signal step completion.
+		if a.workflowEngine != nil {
+			a.workflowEngine.HandleStatusChange(taskID, to)
+		}
+
+		switch to {
+		case string(task.StatusInReview):
+			msg := taskID
+			if t, err := a.tasks.Get(taskID); err == nil {
+				msg = t.Title
+			}
+			a.notifier.Send(notification.LevelInfo, "Ready for review", msg, taskID, "")
+		case string(task.StatusHumanRequired):
+			msg := taskID
+			if t, err := a.tasks.Get(taskID); err == nil {
+				msg = t.Title
+			}
+			a.notifier.Send(notification.LevelWarning, "Needs human", msg, taskID, "")
+			if a.humanReview != nil {
+				go a.humanReview.maybeSpawn(taskID, from)
+			}
+		case string(task.StatusTesting):
+			if a.workflowEngine != nil {
+				if _, err := a.workflowEngine.DispatchEvent(
+					taskID,
+					"task.status_changed",
+					map[string]string{"task.status": string(task.StatusTesting)},
+					nil,
+				); err != nil {
+					a.logger.Error("workflow.dispatch.testing", "task_id", taskID, "err", err)
+				}
+			}
+		}
+	})
+}
+
+func (a *App) initAudit() {
+	al, err := audit.NewLogger(a.auditDir)
+	if err != nil {
+		a.logger.Warn("audit.init.degraded", "err", err)
+		// a.audit remains nil; logAudit() is a no-op when audit is nil.
+		return
+	}
+	a.audit = al
+	retentionDays := a.cfg.Audit.RetentionDays
+	if retentionDays <= 0 {
+		retentionDays = 30
+	}
+	if err := audit.Cleanup(a.auditDir, retentionDays); err != nil {
+		a.logger.Warn("audit.cleanup", "err", err)
+	}
+}
+
+// initProviderHealth constructs the provider health checker, wires it into
+// the agent manager as a gate, and starts its background probe loop. When
+// providers.health_check.enabled=false the checker is skipped entirely and
+// the manager runs with a nil gate (no blocking).
+func (a *App) initProviderHealth(ctx context.Context, emit func(string, any)) {
+	if !a.cfg.Providers.HealthCheck.Enabled {
+		a.logger.Info("provider.health.disabled")
+		return
+	}
+	pc := provider.New(provider.Config{
+		Interval:         time.Duration(a.cfg.Providers.HealthCheck.IntervalSeconds) * time.Second,
+		ClaudeEnabled:    a.cfg.Providers.Claude.Enabled,
+		CodexEnabled:     a.cfg.Providers.Codex.Enabled,
+		AutoFailover:     a.cfg.Providers.AutoFailover,
+		ClaudeRLCooldown: time.Duration(a.cfg.Providers.Claude.RateLimitCooldownSeconds) * time.Second,
+		CodexRLCooldown:  time.Duration(a.cfg.Providers.Codex.RateLimitCooldownSeconds) * time.Second,
+	}, emit, a.logger)
+	a.providerHealth = pc
+	a.agents.SetHealthGate(pc)
+	a.wg.Go(func() { pc.Run(ctx) })
+}
+
+// emitDegradedWarnings fires startup:degraded for any subsystem that failed
+// to initialize. Called after emit is configured so the frontend receives the events.
+func (a *App) emitDegradedWarnings(emit func(string, any)) {
+	type degraded struct {
+		Subsystem string `json:"subsystem"`
+		Reason    string `json:"reason"`
+	}
+	if a.audit == nil {
+		emit(events.StartupDegraded, degraded{"audit", "audit logger failed to initialize; audit trail unavailable"})
+	}
+	if a.stats == nil {
+		emit(events.StartupDegraded, degraded{"stats", "stats store failed to initialize; metrics unavailable"})
+	}
+}
+
+// initAutomations starts every per-machine task source in dependency order
+// and returns the GitHub issues fetcher (still consumed by
+// startBackgroundServices). Extracted so Startup stays under funlen.
+func (a *App) initAutomations(emit func(string, any)) *poll.IssuesFetcher {
+	a.initTodoist(emit)
+	a.initRenovate(emit)
+	a.initTriage()
+	a.initHumanReview()
+	return a.initIssuesFetcher(emit)
+}
+
+func (a *App) initWorkflowEngine() {
+	if os.Getenv("SYBRA_DISABLE_WORKFLOWS") == "1" {
+		a.logger.Info("workflow.disabled")
+		return
+	}
+	wfStore, err := workflow.NewStore(config.WorkflowsDir())
+	if err != nil {
+		a.logger.Error("workflow.store.init", "err", err)
+		return
+	}
+	a.workflowStore = wfStore
+	if syncErr := workflow.SyncBuiltins(wfStore); syncErr != nil {
+		a.logger.Error("workflow.sync-builtins", "err", syncErr)
+	}
+	a.workflowEngine = workflow.NewEngine(
+		wfStore,
+		&taskAdapter{tasks: a.tasks, projects: a.projects},
+		&agentAdapter{agents: a.agents, agentOrch: a.agentOrch, tasks: a.tasks, sandboxes: a.sandboxes},
+		a.logger,
+	)
+	a.workflowEngine.SetPRLinker(prLinkerAdapter{})
+	a.workflowEngine.SetWorktreeGetter(&worktreeGetterAdapter{tasks: a.tasks, mgr: a.worktrees})
+	a.workflowEngine.SetContext(a.ctx)
+	// SetOnComplete moves to wireServices so the callback closure binds
+	// to the AgentCompletionHandler constructed there.
+}
+
+func (a *App) initApprovalServer(emit func(string, any)) {
+	srv, err := agent.NewApprovalServer(emit, a.logger)
+	if err != nil {
+		a.logger.Error("approval-server.init", "err", err)
+		return
+	}
+	srv.SetManager(a.agents)
+	a.agents.SetApprovalAddr(srv.Addr())
+	a.agentSvc.approval = srv
+}
+
+func (a *App) logAudit(eventType, taskID, agentID string, data map[string]any) {
+	if a.audit == nil {
+		return
+	}
+	if err := a.audit.Log(audit.Event{
+		Type:    eventType,
+		TaskID:  taskID,
+		AgentID: agentID,
+		Data:    data,
+	}); err != nil {
+		a.logger.Error("audit.log", "type", eventType, "err", err)
+	}
+}
+
+// seedDefaultLoopAgents creates the built-in sybra-self-monitor loop on
+// first boot only. It is disabled by default so the user can review the
+// configuration in the GUI before enabling. Idempotent: if a record with
+// the same Name already exists this is a no-op.
+func (a *App) initLoopAgents() error {
+	store, err := loopagent.NewStore(a.cfg.LoopAgentsDir)
+	if err != nil {
+		a.logger.Error("loopagent.store.init", "err", err)
+		return err
+	}
+	a.loopAgents = store
+	return nil
+}
+
+func (a *App) initLoopScheduler(ctx context.Context, emit func(string, any)) {
+	a.loopSched = loopagent.NewScheduler(ctx, a.loopAgents, a.agents, a.logger, emit, config.HomeDir())
+	a.seedDefaultLoopAgents()
+	a.loopSched.Sync()
+}
+
+func (a *App) seedDefaultLoopAgents() {
+	if a.loopAgents == nil {
+		return
+	}
+	const name = "sybra-self-monitor"
+	if _, ok := a.loopAgents.FindByName(name); ok {
+		return
+	}
+	created, err := a.loopAgents.Create(loopagent.LoopAgent{
+		Name:         name,
+		Prompt:       "/sybra-self-monitor",
+		IntervalSec:  21600, // 6 hours
+		AllowedTools: []string{"Bash", "Read", "Grep", "Glob"},
+		Provider:     "claude",
+		Model:        "sonnet",
+		Enabled:      false,
+	})
+	if err != nil {
+		a.logger.Warn("loopagent.seed.failed", "name", name, "err", err)
+		return
+	}
+	a.logger.Info("loopagent.seed.created", "id", created.ID, "name", name)
+}
+
+// newRecovery wires the App's deps into a recovery.Recovery used for
+// boot-time cleanup and the periodic restart-stale sweep called from the
+// orchestrator loop. Holds a pointer to a.restartStaleErr so the throttle
+// state is shared across both call sites.
+func (a *App) newRecovery() *recovery.Recovery {
+	var retention time.Duration
+	if days := a.cfg.DefaultLogRetentionDays(); days > 0 {
+		retention = time.Duration(days) * 24 * time.Hour
+	}
+	return &recovery.Recovery{
+		Tasks:          a.tasks,
+		Agents:         a.agents,
+		Worktrees:      a.worktrees,
+		WorkflowEngine: a.workflowEngine,
+		Orchestrator:   a.agentOrch,
+		Projects:       a.projects,
+		Logger:         a.logger,
+		Throttle:       a.restartStaleErr,
+		WG:             &a.wg,
+		LogDir:         a.logDir,
+		LogRetention:   retention,
+	}
+}
+
+// syncSkillsBundle drives the skillsync package with the App's source/dst
+// configuration. UserHomeDir is best-effort — when unavailable the user-home
+// destinations (~/.claude/skills, ~/.codex/skills) are silently skipped so
+// startup still succeeds in environments without a usable home dir.
+func (a *App) syncSkillsBundle() {
+	userHome, err := os.UserHomeDir()
+	if err != nil {
+		a.logger.Debug("skills.sync.no_user_home", "err", err)
+		userHome = ""
+	}
+	(&skillsync.Syncer{Logger: a.logger}).Run(skillsync.Options{
+		RepoDir:      a.repoDir,
+		SkillsFS:     a.skillsFS,
+		PrimaryDst:   a.skillsDir,
+		SybraHomeDir: config.HomeDir(),
+		UserHomeDir:  userHome,
+	})
+}
